+++ lib/src/track/local/local.dart
@@ -80,7 +80,12 @@ abstract class LocalTrack extends Track {
           source,
           mediaStream,
           mediaStreamTrack,
-        );
+        ) {
+    mediaStreamTrack.onEnded = () {
+      logger.fine('MediaStreamTrack.onEnded()');
+      events.emit(TrackEndedEvent(track: this));
+    };
+  }
 
   /// Mutes this [LocalTrack]. This will stop the sending of track data
   /// and notify the [RemoteParticipant] with [TrackMutedEvent].
@@ -153,6 +158,13 @@ abstract class LocalTrack extends Track {
         }
         if (options.selfBrowserSurface != null) {
           constraints['selfBrowserSurface'] = options.selfBrowserSurface!;
+        }
+
+        // Remove resolution settings to fix low-resolution screen share on Safari 17.
+        // related bug: https://bugs.webkit.org/show_bug.cgi?id=263015
+        if (lkBrowser() == BrowserType.safari &&
+            lkBrowserVersion().major == 17) {
+          constraints['video'] = true;
         }
       }
       stream = await rtc.navigator.mediaDevices.getDisplayMedia(constraints);