+++ lib/src/utils.dart
@@ -271,6 +271,7 @@ class Utils {
     required bool isScreenShare,
     required VideoDimensions dimensions,
     required List<VideoParameters> presets,
+    String? codec,
   }) {
     assert(presets.isNotEmpty, 'presets should not be empty');
     VideoEncoding result = presets.first.encoding;
@@ -281,6 +282,26 @@ class Utils {
     for (final preset in presets) {
       result = preset.encoding;
       if (preset.dimensions.width >= size) break;
+    }
+
+    // presets are based on the assumption of vp8 as a codec
+    // for other codecs we adjust the maxBitrate if no specific videoEncoding has been provided
+    // users should override these with ones that are optimized for their use case
+    // NOTE: SVC codec bitrates are inclusive of all scalability layers. while
+    // bitrate for non-SVC codecs does not include other simulcast layers.
+    if (codec != null) {
+      switch (codec) {
+        case 'av1':
+          result =
+              result.copyWith(maxBitrate: (result.maxBitrate * 0.7).toInt());
+          break;
+        case 'vp9':
+          result =
+              result.copyWith(maxBitrate: (result.maxBitrate * 0.85).toInt());
+          break;
+        default:
+          break;
+      }
     }
 
     return result;
@@ -371,6 +392,7 @@ class Utils {
     required bool isScreenShare,
     VideoDimensions? dimensions,
     VideoPublishOptions? options,
+    String? codec,
   }) {
     options ??= const VideoPublishOptions();
 
@@ -397,6 +419,7 @@ class Utils {
         isScreenShare: isScreenShare,
         dimensions: dimensions,
         presets: presets,
+        codec: codec,
       );
 
       logger.fine('using video encoding', videoEncoding);
@@ -410,20 +433,22 @@ class Utils {
     if (scalabilityMode != null && isSVCCodec(options.videoCodec)) {
       logger.info('using svc with scalabilityMode ${scalabilityMode}');
 
-      final sm = ScalabilityMode(scalabilityMode);
+      //final sm = ScalabilityMode(scalabilityMode);
 
-      List<rtc.RTCRtpEncoding> encodings = [];
-
+      List<rtc.RTCRtpEncoding> encodings = [videoEncoding.toRTCRtpEncoding()];
+      /*
       if (sm.spatial > 3) {
         throw Exception('unsupported scalabilityMode: ${scalabilityMode}');
       }
       for (int i = 0; i < sm.spatial; i += 1) {
         encodings.add(rtc.RTCRtpEncoding(
-          rid: videoRids[i],
-          maxBitrate: (videoEncoding.maxBitrate / 3 * (i + 1)).toInt(),
+          rid: videoRids[2 - i],
+          maxBitrate: videoEncoding.maxBitrate ~/ math.pow(3, i),
           maxFramerate: videoEncoding.maxFramerate,
+          scaleResolutionDownBy: null,
+          numTemporalLayers: sm.temporal.toInt(),
         ));
-      }
+      }*/
       encodings[0].scalabilityMode = scalabilityMode;
       logger.fine('encodings $encodings');
       return encodings;
@@ -432,6 +457,7 @@ class Utils {
       return [videoEncoding.toRTCRtpEncoding()];
     }
 
+    // compute simulcast encodings
     final userParams = isScreenShare
         ? options.screenShareSimulcastLayers
         : options.videoSimulcastLayers;
@@ -466,12 +492,13 @@ class Utils {
   @internal
   static List<rtc.RTCRtpEncoding>? computeTrackBackupEncodings(
     LocalVideoTrack track,
-    VideoPublishOptions opts,
+    BackupVideoCodec backupOpts,
   ) {
-    opts = opts.copyWith(
-        videoCodec: opts.backupCodec!.codec,
-        videoEncoding: opts.backupCodec!.encoding);
-
+    final opts = VideoPublishOptions(
+      videoCodec: backupOpts.codec,
+      videoEncoding: backupOpts.encoding,
+      simulcast: backupOpts.simulcast,
+    );
     var encodings = computeVideoEncodings(
       isScreenShare: track.source == TrackSource.screenShareVideo,
       dimensions: track.currentOptions.params.dimensions,
@@ -484,6 +511,7 @@ class Utils {
   static List<lk_models.VideoLayer> computeVideoLayers(
     VideoDimensions dimensions,
     List<rtc.RTCRtpEncoding>? encodings,
+    bool isSVC,
   ) {
     // default to a single layer, HQ
     if (encodings == null) {
@@ -495,6 +523,22 @@ class Utils {
           bitrate: 0,
         )
       ];
+    }
+
+    if (isSVC) {
+      final sm = ScalabilityMode(encodings[0].scalabilityMode ?? 'L3T3_KEY');
+      final List<lk_models.VideoLayer> layers = [];
+      final maxBitrate = encodings[0].maxBitrate ?? 0;
+      for (var i = 0; i < sm.spatial; i++) {
+        layers.add(lk_models.VideoLayer(
+          quality: lk_models.VideoQuality.valueOf(
+              lk_models.VideoQuality.HIGH.value - i),
+          width: (dimensions.width / math.pow(2, i)).floor(),
+          height: (dimensions.height / math.pow(2, i)).floor(),
+          bitrate: (maxBitrate / math.pow(3, i)).ceil(),
+        ));
+      }
+      return layers;
     }
 
     return encodings.map((e) {
@@ -575,4 +619,15 @@ class ScalabilityMode {
   String toString() {
     return 'L${spatial}T${temporal}${suffix ?? ''}';
   }
+}
+
+String mimeTypeToVideoCodecString(String mimeType) {
+  if (!mimeType.contains('/') && mimeType.split('/').length != 2) {
+    throw Exception('Invalid mimeType: $mimeType');
+  }
+  final codec = mimeType.split('/')[1].toLowerCase();
+  if (!videoCodecs.contains(codec)) {
+    throw Exception('Video codec not supported: $codec');
+  }
+  return codec;
 }