+++ lib/src/widgets/video_track_renderer.dart
@@ -19,11 +19,13 @@ import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
 import '../events.dart';
 import '../extensions.dart';
 import '../internal/events.dart';
+import '../logger.dart';
 import '../managers/event.dart';
 import '../support/platform.dart';
 import '../track/local/local.dart';
 import '../track/local/video.dart';
 import '../track/options.dart';
+import '../types/other.dart';
 
 enum VideoViewMirrorMode {
   auto,
@@ -49,40 +51,49 @@ class VideoTrackRenderer extends StatefulWidget {
 }
 
 class _VideoTrackRendererState extends State<VideoTrackRenderer> {
-  final _renderer = rtc.RTCVideoRenderer();
-  bool _rendererReady = false;
+  rtc.RTCVideoRenderer? _renderer;
   EventsListener<TrackEvent>? _listener;
   // Used to compute visibility information
   late GlobalKey _internalKey;
+
+  Future<rtc.RTCVideoRenderer?> _initializeRenderer() async {
+    _renderer ??= rtc.RTCVideoRenderer();
+    await _renderer!.initialize();
+    await _attach();
+    return _renderer;
+  }
+
+  void disposeRenderer() {
+    try {
+      _renderer?.srcObject = null;
+      _renderer?.dispose();
+      _renderer = null;
+    } catch (e) {
+      logger.warning('Got error disposing renderer: $e');
+    }
+  }
 
   @override
   void initState() {
     super.initState();
     _internalKey = widget.track.addViewKey();
-
-    (() async {
-      await _renderer.initialize();
-      await _attach();
-      setState(() => _rendererReady = true);
-    })();
   }
 
   @override
   void dispose() {
     widget.track.removeViewKey(_internalKey);
     _listener?.dispose();
-    _renderer.srcObject = null;
-    _renderer.dispose();
+    disposeRenderer();
     super.dispose();
   }
 
   Future<void> _attach() async {
-    _renderer.srcObject = widget.track.mediaStream;
+    _renderer?.srcObject = widget.track.mediaStream;
     await _listener?.dispose();
     _listener = widget.track.createListener()
       ..on<TrackStreamUpdatedEvent>((event) {
         if (!mounted) return;
-        _renderer.srcObject = event.stream;
+        _renderer?.srcObject = event.stream;
       })
       ..on<LocalTrackOptionsUpdatedEvent>((event) {
         if (!mounted) return;
@@ -104,31 +115,39 @@ class _VideoTrackRendererState extends State<VideoTrackRenderer> {
 
     if ([BrowserType.safari, BrowserType.firefox].contains(lkBrowser()) &&
         oldWidget.key != widget.key) {
-      _renderer.srcObject = widget.track.mediaStream;
+      _renderer?.srcObject = widget.track.mediaStream;
     }
   }
 
   @override
-  Widget build(BuildContext context) => !_rendererReady
-      ? Container()
-      : Builder(
-          key: _internalKey,
-          builder: (ctx) {
-            // let it render before notifying build
-            WidgetsBindingCompatible.instance
-                ?.addPostFrameCallback((timeStamp) {
-              widget.track.onVideoViewBuild?.call(_internalKey);
-            });
-            return rtc.RTCVideoView(
-              _renderer,
-              mirror: _shouldMirror(),
-              filterQuality: FilterQuality.medium,
-              objectFit: widget.fit,
-            );
-          },
-        );
+  Widget build(BuildContext context) => FutureBuilder(
+      future: _initializeRenderer(),
+      builder: (context, snapshot) {
+        if (snapshot.hasData && _renderer != null) {
+          return Builder(
+            key: _internalKey,
+            builder: (ctx) {
+              // let it render before notifying build
+              WidgetsBindingCompatible.instance
+                  ?.addPostFrameCallback((timeStamp) {
+                widget.track.onVideoViewBuild?.call(_internalKey);
+              });
+              return rtc.RTCVideoView(
+                _renderer!,
+                mirror: _shouldMirror(),
+                filterQuality: FilterQuality.medium,
+                objectFit: widget.fit,
+              );
+            },
+          );
+        }
+
+        return Container();
+      });
 
   bool _shouldMirror() {
+    // off for screen share
+    if (widget.track.source == TrackSource.screenShareVideo) return false;
     // on
     if (widget.mirrorMode == VideoViewMirrorMode.mirror) return true;
     // auto